#filmorate refactor перенес исключения в другой пакет
