+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/InvalidFilmException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.exceptions;
 
 public class InvalidFilmException extends RuntimeException {
     public InvalidFilmException(String message) {
